add data validation

+++ Controllers/StockController.cs
@@ -1,6 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
 
-using api.Data;
 using api.Mappers;
 using api.Dtos.Stock;
 using api.Interfaces;
@@ -11,65 +10,76 @@ namespace api.Controllers
   [ApiController]
   public class StockController : ControllerBase
   {
-    private readonly ApplicationDBContext _context;
     private readonly IStockRepository _stockRepo;
-    public StockController(ApplicationDBContext context, IStockRepository stockRepo)
+    public StockController(IStockRepository stockRepo)
     {
       _stockRepo = stockRepo;
-      _context = context;
     }
 
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+      if (!ModelState.IsValid)
+        return BadRequest(ModelState);
+
       var stocks = await _stockRepo.GetAllAsync();
       var stockDto = stocks.Select(s => s.ToStockDto());
+
       return Ok(stockDto);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+      if (!ModelState.IsValid)
+        return BadRequest(ModelState);
+
       var stock = await _stockRepo.GetByIdAsync(id);
+
       if (null == stock)
-      {
         return NotFound();
-      }
+
       return Ok(stock.ToStockDto());
     }
 
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
     {
+      if (!ModelState.IsValid)
+        return BadRequest(ModelState);
+
       var stockModel = stockDto.ToStockFromCreateDTO();
       await _stockRepo.CreateAsync(stockModel);
+
       return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
     }
 
     [HttpPut]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
     {
+      if (!ModelState.IsValid)
+        return BadRequest(ModelState);
+
       var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
       if (null == stockModel)
-      {
         return NotFound();
-      }
 
       return Ok(stockModel.ToStockDto());
     }
 
     [HttpDelete]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+      if (!ModelState.IsValid)
+        return BadRequest(ModelState);
+
       var stockModel = await _stockRepo.DeleteAsync(id);
 
       if (null == stockModel)
-      {
         return NotFound();
-      }
 
       return NoContent();
     }
